LIX-263 # features library refactoring, styling and bug fixing
