+++ packages/lixpi/constants/ts/types.ts
@@ -1195,9 +1195,16 @@ export type ExtractionRun = {
     userId: string
     status: ExtractionRunStatus
     featureId?: string
+    userText?: string
+    modelConfig?: {
+        analysisModelId?: string
+        mediaModelId?: string
+    }
     transcriptJson?: object
     sourceContextSnapshot?: object
     trace?: StageTraceEvent[]
+    stageReasoning?: Record<string, string>
+    featureCard?: Record<string, any>
     error?: string
     createdAt: number
     updatedAt: number
@@ -1230,9 +1237,14 @@ export type CanvasAiChatPanelState = {
 
 export type CanvasFeatureExtractionState = {
     extractionRunId: string
+    featureId?: string
     status: ExtractionRunStatus
     userText?: string
     aiProvider?: string
+    modelConfig?: {
+        analysisModelId?: string
+        mediaModelId?: string
+    }
     stepDetails?: Record<string, string>
     reasoningText?: string
     // Streamed model output (thinking/reasoning) keyed by the stage that produced it,
@@ -1253,6 +1265,9 @@ export type CanvasState = {
     aiChatSidebarTabs?: CanvasAiChatSidebarTab[]
     activeAiChatSidebarTabId?: string
     aiChatPanel?: CanvasAiChatPanelState
+    // Legacy workspace-state field. Current feature extraction state is owned by
+    // API ExtractionRun records; clients may read this only to prune old pending
+    // placeholders from saved workspaces.
     featureExtractionRuns?: Record<string, CanvasFeatureExtractionState>
 }
