+++ services/web-ui/src/components/aiModelControls/index.ts
@@ -12,7 +12,7 @@ export {
     type AiModelDropdownOption,
     type MediaGenerationConfigMatrixControls,
     type MediaGenerationConfigMatrixViewInstance,
-} from '$src/components/aiModelControls/index.ts'
+} from '$src/components/aiModelControls/aiModelControls.ts'
 
 export {
     createGenericAiModelMultiSelect,
@@ -21,4 +21,13 @@ export {
     type AiModelMultiSelectControls,
     type ImageModelMultiSelectControls,
     type VideoModelMultiSelectControls,
-} from '$src/components/aiModelControls/index.ts'
+} from '$src/components/aiModelControls/modelMultiSelect.ts'
+
+export {
+    applyAiModelMenuStyleSettings,
+    createAiModelMenuContent,
+    type AiModelMenuContentView,
+    type AiModelMenuControlItem,
+    type AiModelMenuSectionConfig,
+    type AiModelMenuSectionView,
+} from '$src/components/aiModelControls/modelMenuSection.ts'
