+++ services/web-ui/src/components/proseMirror/plugins/slashCommandsMenuPlugin/README.md
@@ -71,6 +71,8 @@ Uses ProseMirror's `coordsAtPos()` to get screen coordinates at the `/` trigger 
 | Image | image, img, picture | Insert an image |
 | Table | table | Insert a table (Phase 2) |
 | File | file, attachment | Attach a file (Phase 2) |
+| Use Feature | use, feature, features | Open the Features panel to pick a reusable feature |
+| Extract Feature | extract, extract-feature | Open the Features panel with a pending extraction confirmation |
 
 ## Adding New Commands
